# StudentVersionManagement/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/delete_fatal_peer": {
            "post": {
                "description": "删除故障节点",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Raft管理"
                ],
                "summary": "删除故障节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "故障节点ID",
                        "name": "PeerID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "故障节点地址",
                        "name": "PeerAddress",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "故障节点端口地址",
                        "name": "PeerPortAddress",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功删除节点"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    }
                }
            }
        },
        "/get_leader_port_address": {
            "get": {
                "description": "获取领导者端口的地址",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Raft管理"
                ],
                "summary": "获取领导者端口地址",
                "responses": {
                    "200": {
                        "description": "成功获取领导者端口地址"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    }
                }
            }
        },
        "/join_raft_cluster": {
            "post": {
                "description": "将自身加入到Raft集群中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Raft管理"
                ],
                "summary": "加入Raft集群",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "nodeID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "节点地址",
                        "name": "nodeAddress",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "端口地址",
                        "name": "portAddress",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加节点成功"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    }
                }
            }
        },
        "/leader_handle_command": {
            "post": {
                "description": "处理领导者节点发送的命令",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Raft管理"
                ],
                "summary": "处理领导者命令",
                "parameters": [
                    {
                        "type": "string",
                        "description": "命令数据",
                        "name": "cmd",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "领导者节点已处理命令"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    }
                }
            }
        },
        "/versions": {
            "post": {
                "description": "添加一个新的版本号信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "版本管理"
                ],
                "summary": "添加版本号信息",
                "parameters": [
                    {
                        "description": "版本号信息",
                        "name": "version",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Version"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功添加版本号"
                    },
                    "400": {
                        "description": "无效的请求体"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    }
                }
            }
        },
        "/versions/{id}": {
            "get": {
                "description": "根据ID获取版本号信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "版本管理"
                ],
                "summary": "获取指定ID的版本号信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "版本号ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取版本号",
                        "schema": {
                            "$ref": "#/definitions/model.Version"
                        }
                    },
                    "404": {
                        "description": "版本号未找到"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    }
                }
            },
            "put": {
                "description": "根据ID更新版本号信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "版本管理"
                ],
                "summary": "更新版本号信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "版本号ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "版本号信息",
                        "name": "version",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Version"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "版本号更新成功"
                    },
                    "400": {
                        "description": "无效的请求体或版本号ID"
                    },
                    "404": {
                        "description": "版本号未找到"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    }
                }
            },
            "delete": {
                "description": "根据ID删除版本号信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "版本管理"
                ],
                "summary": "删除版本号信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "版本号ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "版本号删除成功"
                    },
                    "400": {
                        "description": "无效的版本号ID"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Version": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID 是版本的唯一标识符，用作主键。",
                    "type": "string"
                },
                "platform": {
                    "description": "Platform 表示版本适用的平台，不能为空。",
                    "type": "string"
                },
                "versionNo": {
                    "description": "VersionNo 表示版本号，不能为空。",
                    "type": "string"
                }
            }
        }
    }
}